requirements.txt
aws-cdk-lib==2.45.0
constructs>=10.0.0,<11.0.0
cdk-nag>=0.1.7
jsonify
requests
argparse
pytz
boto3
